+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 GAMSPy CHANGELOG
 ================
 
+GAMSPy 0.13.4
+-------------
+- General
+  - Fix hanging issue on Windows for GAMS Engine backend.
+  - Refactor toGams converter.
+  - Fix solver options file path bug.
+- Testing
+  - Add more tests for GAMS MIRO.
+
 GAMSPy 0.13.3
 -------------
 - General